tweak
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@458 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_ivr.c
@@ -89,7 +89,7 @@ SWITCH_DECLARE(switch_status) switch_ivr_record_file(switch_core_session *sessio
 	
 
 	while (switch_channel_get_state(channel) == CS_EXECUTE) {
-		int len;
+		size_t len;
 
 		if (dtmf_callback) {
 			/*
@@ -110,7 +110,7 @@ SWITCH_DECLARE(switch_status) switch_ivr_record_file(switch_core_session *sessio
 			break;
 		}
 
-		len = read_frame->datalen / 2;
+		len = (size_t) read_frame->datalen / 2;
 		switch_core_file_write(&fh, read_frame->data, &len);
 	}
